+++ src/cone/app/tests/__init__.py
@@ -62,7 +62,7 @@ def test_suite():  # pragma: no cover
     suite.addTest(unittest.findTestCases(test_browser_login))
     suite.addTest(unittest.findTestCases(test_browser_order))
     suite.addTest(unittest.findTestCases(test_browser_referencebrowser))
-    # suite.addTest(unittest.findTestCases(test_browser_resources)) # XXX: can wait until resources finalized
+    suite.addTest(unittest.findTestCases(test_browser_resources))
     suite.addTest(unittest.findTestCases(test_browser_search))
     suite.addTest(unittest.findTestCases(test_browser_settings))
     suite.addTest(unittest.findTestCases(test_browser_sharing))